web/pyproject.toml
[project]
name = "agent-web"
version = "0.1.0"
description = "NiceGUI frontend for FastAPI LangGraph Agent"
requires-python = ">=3.13"
dependencies = [
    "nicegui>=2.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "markdown>=3.5",
]
